# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "4Foods API",
    "description": "API documentation for 4Foods Android app",
    "version": "1.0.0"
  },
  "host": "api.4foods.app",
  "basePath": "/",
  "schemes": ["https"],
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "Enter your bearer token in the format **Bearer <token>**"
    }
  },
  "paths": {
...
    "/products": {
      "post": {
        "tags": ["Product"],
        "summary": "Create product (seller or admin)",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "required": true,
            "schema": {
              "type": "object",
              "required": ["name", "price", "categoryId", "imageUrl"],
              "properties": {
                "name": { "type": "string" },
                "price": { "type": "number" },
                "description": { "type": "string" },
                "categoryId": { "type": "string" },
                "imageUrl": { "type": "string" },
                "discountPercent": { "type": "number" },
                "prepTime": { "type": "number" },
                "stock": { "type": "number" },
                "shopId": {
                  "type": "string",
                  "description": "Admin only (override seller shop)"
                }
              }
            },
            "example": {
              "name": "Trà Sữa Truyền Thống",
              "price": 30000,
              "description": "Vị truyền thống thơm béo",
              "categoryId": "671acb0298f2c700f9e2c5a9",
              "imageUrl": "https://cdn.4foods.app/p/tra-sua.jpg",
              "discountPercent": 10,
              "prepTime": 5,
              "stock": 120
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "examples": {
              "application/json": {
                "message": "Product created & pending approval",
                "product": {
                  "_id": "67310df7d9c6a8f016ca9123",
                  "name": "Trà Sữa Truyền Thống",
                  "price": 30000,
                  "status": "pending",
                  "category": "671acb0298f2c700f9e2c5a9",
                  "discountPercent": 10,
                  "stock": 120,
                  "createdAt": "2025-11-20T08:10:00.000Z"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "examples": {
              "application/json": { "error": "Missing required fields" }
            }
          },
          "403": {
            "description": "Forbidden",
            "examples": {
              "application/json": {
                "error": "Only sellers or admins can add products"
              }
            }
          }
        }
      },
      "get": {
        "tags": ["Product"],
        "summary": "Search & list products (normalized, filter, sort, paginate)",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "shopId",
            "in": "query",
            "type": "string",
            "required": false,
            "example": "671a6cb29e2c5a9d"
          },
          {
            "name": "categoryId",
            "in": "query",
            "type": "string",
            "required": false,
            "example": "671acb0298f2c700f9e2c5a9"
          },
          {
            "name": "keyword",
            "in": "query",
            "type": "string",
            "required": false,
            "example": "tra sua truyen thong"
          },
          {
            "name": "sort",
            "in": "query",
            "type": "string",
            "required": false,
            "enum": ["rate", "price", "recent"],
            "example": "rate"
          },
          {
            "name": "page",
            "in": "query",
            "type": "integer",
            "required": false,
            "example": 1
          },
          {
            "name": "limit",
            "in": "query",
            "type": "integer",
            "required": false,
            "example": 10
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list",
            "examples": {
              "application/json": {
                "total": 57,
                "page": 1,
                "totalPages": 6,
                "products": [
                  {
                    "_id": "67310df7d9c6a8f016ca9123",
                    "name": "Trà Sữa Truyền Thống",
                    "imageUrl": "https://cdn.4foods.app/p/tra-sua.jpg",
                    "price": 30000,
                    "stock": 120,
                    "rating": 4.7,
                    "discountPercent": 10,
                    "description": "Vị truyền thống thơm béo",
                    "createdAt": "2025-11-20T08:10:00.000Z",
                    "category": "Trà sữa",
                    "shop": "Tiệm Trà Sữa Posu",
                    "shopAddress": "123 Đường Hàn Thuyên"
                  }
                ]
              }
            }
          },
          "500": {
            "description": "Server error",
            "examples": {
              "application/json": { "error": "Internal server error" }
            }
          }
        }
      }
    },
    "/products/{id}": {
      "get": {
        "tags": ["Product"],
        "summary": "Get product details (admin only)",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "67310df7d9c6a8f016ca9123"
          }
        ],
        "responses": {
          "200": {
            "description": "Details",
            "examples": {
              "application/json": {
                "_id": "67310df7d9c6a8f016ca9123",
                "name": "Trà Sữa Truyền Thống",
                "price": 30000,
                "status": "pending",
                "category": { "name": "Trà sữa" },
                "shopId": {
                  "name": "Tiệm Trà Sữa Posu",
                  "address": "123 Đường Hàn Thuyên",
                  "isOpen": true
                },
                "sellerId": {
                  "name": "Posu Seller",
                  "email": "seller@4foods.app"
                },
                "comments": []
              }
            }
          },
          "404": {
            "description": "Not found",
            "examples": { "application/json": { "error": "Product not found" } }
          },
          "500": { "description": "Server error" }
        }
      },
      "put": {
        "tags": ["Product"],
        "summary": "Update product (seller own or admin)",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "67310df7d9c6a8f016ca9123"
          },
          {
            "in": "body",
            "name": "body",
            "schema": { "type": "object" },
            "example": {
              "name": "Trà Sữa Truyền Thống (Mới)",
              "price": 32000,
              "discountPercent": 5
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Updated",
            "examples": {
              "application/json": {
                "_id": "67310df7d9c6a8f016ca9123",
                "status": "pending"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "examples": {
              "application/json": {
                "error": "You can only update your own products"
              }
            }
          },
          "400": {
            "description": "Invalid",
            "examples": {
              "application/json": {
                "error": "Pending product cannot be edited"
              }
            }
          },
          "404": { "description": "Not found" }
        }
      },
      "delete": {
        "tags": ["Product"],
        "summary": "Delete product (seller own or admin)",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "67310df7d9c6a8f016ca9123"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "examples": {
              "application/json": { "message": "Product deleted successfully" }
            }
          },
          "403": {
            "description": "Forbidden",
            "examples": {
              "application/json": {
                "error": "You can only delete your own product"
              }
            }
          },
          "404": { "description": "Not found" }
        }
      }
    },
    "/products/{id}/approve": {
      "put": {
        "tags": ["Product"],
        "summary": "Admin approve/reject product",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "67310df7d9c6a8f016ca9123"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "required": ["status"],
              "properties": {
                "status": {
                  "type": "string",
                  "enum": ["displayed", "rejected", "pending"]
                },
                "reason": { "type": "string" }
              }
            },
            "example": { "status": "displayed" }
          }
        ],
        "responses": {
          "200": {
            "description": "Status updated",
            "examples": {
              "application/json": {
                "message": "Product status updated to displayed"
              }
            }
          },
          "400": {
            "description": "Invalid status",
            "examples": { "application/json": { "error": "Invalid status" } }
          },
          "403": {
            "description": "Forbidden",
            "examples": { "application/json": { "error": "Admin only" } }
          },
          "404": { "description": "Not found" }
        }
      }
    },
    "/products/{id}/toggle": {
      "put": {
        "tags": ["Product"],
        "summary": "Seller toggle displayed/hidden",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "67310df7d9c6a8f016ca9123"
          }
        ],
        "responses": {
          "200": {
            "description": "Toggled",
            "examples": {
              "application/json": { "message": "Product is now hidden" }
            }
          },
          "400": {
            "description": "Invalid",
            "examples": {
              "application/json": {
                "error": "Only approved products can be toggled"
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "examples": { "application/json": { "error": "Forbidden" } }
          },
          "404": { "description": "Not found" }
        }
      }
    },
    "/products/{id}/comment": {
      "post": {
        "tags": ["Product"],
        "summary": "Add comment",
        "security": [{ "Bearer": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "67310df7d9c6a8f016ca9123"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "object",
              "required": ["content"],
              "properties": {
                "content": { "type": "string" },
                "rating": { "type": "number" },
                "images": { "type": "array", "items": { "type": "string" } }
              }
            },
            "example": {
              "content": "Ngon và mát!",
              "rating": 5,
              "images": ["https://cdn.4foods.app/comments/c1.jpg"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Added",
            "examples": {
              "application/json": {
                "message": "Comment added",
                "comment": { "content": "Ngon và mát!", "rating": 5 }
              }
            }
          },
          "404": {
            "description": "Not found",
            "examples": { "application/json": { "error": "Product not found" } }
          },
          "400": {
            "description": "Validation",
            "examples": {
              "application/json": { "error": "Comment content required" }
            }
          }
        }
      }
    },
    "/products/{id}/comments": {
      "get": {
        "tags": ["Product"],
        "summary": "List comments",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "example": "67310df7d9c6a8f016ca9123"
          }
        ],
        "responses": {
          "200": {
            "description": "Comments",
            "examples": {
              "application/json": [
                {
                  "user": { "name": "Minh Tran", "avatar": "/avatars/1.png" },
                  "content": "Rất ngon!",
                  "rating": 5,
                  "images": []
                }
              ]
            }
          },
          "404": {
            "description": "Not found",
            "examples": { "application/json": { "error": "Product not found" } }
          }
        }
      }
    },
...
  }
}